requirements.txt
# LLMs
langchain==0.0.223
# git+https://github.com/hwchase17/langchain.git@master#egg=langchain
openai~=0.27.8

# Guardrails
guardrails-ai==0.1.8
faiss-cpu

# Streamlit
streamlit==1.25.0
streamlit-elements==0.1.0
# streamlit_cookies_manager==0.2.0
extra_streamlit_components==0.1.56

# Google Cloud Platform
httpx_oauth==0.13.0
google-cloud-firestore==2.11.1
google-cloud-bigquery[pandas]==3.11.1
google-api-python-client==2.70.0
google-auth-httplib2
google-auth-oauthlib
firebase
firebase-admin
fireo

# Data science
plotly~=5.16.0
pandas~=2.0.3
scipy
scikit-learn
numpy~=1.25.2
statsmodels
dataframe_image
matplotlib

# Miscellaneous

# altair
watchdog

# wolframalpha
python-dotenv~=1.0.0
# google-search-results
ipykernel==6.25.1
# jupyterlab
pytest~=7.4.0
pytest-xdist
# notebook
# ipywidgets
pandas
numpy
# pyarrow
# PyPDF2
unstructured
# chromadb
lxml
tiktoken
# pinecone-client
# faiss-cpu
deeplake
networkx
pg8000
pulumi~=3.78.1
pulumi_gcp
gitignore_parser
sentry_sdk
flask_sqlalchemy
sqlalchemy-bigquery[bqstorage]
sqlalchemy>=2.0.13
nbformat
sqlparse~=0.4.4
google_cloud_api_keys

# API
openapi-generator
gunicorn
connexion[swagger-ui]==2.14.2
api/clients/python_client/
flask>=2.0.3
flask-cors==4.0.0
authlib>=1.0
typeguard==4.1.3

# Caching
gptcache
# redis
# redis_om
# transformers
# onnxruntime
# tensorflow==2.13.0
torch

# FastAPI
# fastapi==0.45.0
# uvicorn
# pydantic>=1.2.0,<2.0.0
# pydantic_settings

python-jose
six
st_pages==0.4.4
RestrictedPython==6.1

# All bots
kaleido
frozendict

# Discord bot
discord~=2.3.2

# Slack bot
slack_bolt

# Telegram bot
python-telegram-bot

# Code linting and formatting
flake8
isort==5.12.0
black
